+++ package.json
@@ -8,13 +8,13 @@
     "start:prod": "node lib/index.js",
     "demo": "nodemon --demo",
     "ganache": "npx ganache-cli",
-    "test": "npx jest src/*/* --verbose --coverage --runInBand",
-    "test:ci": "npx jest src/*/* --coverage --ci --runInBand",
-    "test:watch": "npx jest src/*/* --verbose --coverage --watch --runInBand",
+    "test": "npx jest test/*/* --verbose --coverage --runInBand",
+    "test:ci": "npx jest test/*/* --coverage --ci --runInBand",
+    "test:watch": "npx jest test/*/* --verbose --coverage --watch --runInBand",
     "build": "rimraf ./lib && npx tsc",
     "build:prod": "rimraf ./lib && tsc -p tsconfig.json",
-    "lint": "npx eslint ./src --ext .json,.js,.ts",
-    "lint:fix": "npx eslint ./src --ext .json,.js,.ts --fix"
+    "lint": "npx eslint ./src ./test --ext .json,.js,.ts",
+    "lint:fix": "npx eslint ./src ./test --ext .json,.js,.ts --fix"
   },
   "repository": {
     "type": "git",
