+++ test/Core-WebSocket.test.ts
@@ -1,5 +1,5 @@
 import { BLOCKCHAIN_WS_URL } from './constants'
-import { WebSocketListener } from '../model'
+import { WebSocketListener } from '../src/model'
 import { runCoreWith } from './CoreRunner'
 
 runCoreWith('WebSocketListener', WebSocketListener, BLOCKCHAIN_WS_URL)
