+++ test/CoreRunner.ts
@@ -1,16 +1,15 @@
-import { createDbConnection } from '../storage/createDbConnection'
 import { deleteDatabase, resetDatabase, sleep } from './utils'
-import { deployAllContracts, ISetup, setupContracts } from '../scripts'
+import { deployAllContracts, ISetup, setupContracts } from '../src/scripts'
 import { BLOCKCHAIN_HTTP_URL } from './constants'
 import { Connection, Repository } from 'typeorm'
-import { ScheduledTransaction, EMetatransactionState } from '../entities'
-import { Cache, Store } from '../storage'
+import { ScheduledTransaction, EMetatransactionState } from '../src/entities'
+import { Cache, Store, createDbConnection } from '../src/storage'
 import { addMinutes } from 'date-fns'
 import Web3 from 'web3'
-import { Recoverer, Collector } from '../model'
-import Core from '../Core'
+import { Recoverer, Collector } from '../src/model'
+import Core from '../src/Core'
 import { ExecutorMock, SchedulerMock } from './mocks'
-import { BlockchainDate } from '../time'
+import { BlockchainDate } from '../src/time'
 import { time } from '@openzeppelin/test-helpers'
 import tracer from 'tracer'
