+++ test/Listener-Polling.test.ts
@@ -1,5 +1,5 @@
 import { BLOCKCHAIN_HTTP_URL } from './constants'
-import { PollingListener } from '../model'
+import { PollingListener } from '../src/model'
 import { runListenerWith } from './ListenerRunner'
 
 runListenerWith('PollingListener', PollingListener, BLOCKCHAIN_HTTP_URL)
