+++ test/Listener-WebSocket.test.ts
@@ -1,5 +1,5 @@
 import { BLOCKCHAIN_WS_URL } from './constants'
-import { WebSocketListener } from '../model'
+import { WebSocketListener } from '../src/model'
 import { runListenerWith } from './ListenerRunner'
 
 runListenerWith('WebSocketListener', WebSocketListener, BLOCKCHAIN_WS_URL)
